Moved common code to VectorSimilarityValuesSource and addressed other comments
