+++ lucene/core/src/java/org/apache/lucene/search/ByteVectorSimilarityValuesSource.java
@@ -28,13 +28,12 @@ import org.apache.lucene.index.VectorSimilarityFunction;
  * A {@link DoubleValuesSource} which computes the vector similarity scores between the query vector
  * and the {@link org.apache.lucene.document.KnnByteVectorField} for documents.
  */
-class ByteVectorSimilarityValuesSource extends DoubleValuesSource {
+class ByteVectorSimilarityValuesSource extends VectorSimilarityValuesSource {
   private final byte[] queryVector;
-  private final String fieldName;
 
   public ByteVectorSimilarityValuesSource(byte[] vector, String fieldName) {
+    super(fieldName);
     this.queryVector = vector;
-    this.fieldName = fieldName;
   }
 
   @Override
@@ -50,24 +49,10 @@ class ByteVectorSimilarityValuesSource extends DoubleValuesSource {
 
       @Override
       public boolean advanceExact(int doc) throws IOException {
-        if (doc >= vectorValues.docID()
-            && (vectorValues.docID() == doc || vectorValues.advance(doc) == doc)) {
-          return true;
-        } else {
-          return false;
-        }
+        return doc >= vectorValues.docID()
+            && (vectorValues.docID() == doc || vectorValues.advance(doc) == doc);
       }
     };
-  }
-
-  @Override
-  public boolean needsScores() {
-    return false;
-  }
-
-  @Override
-  public DoubleValuesSource rewrite(IndexSearcher reader) throws IOException {
-    return this;
   }
 
   @Override
@@ -91,10 +76,5 @@ class ByteVectorSimilarityValuesSource extends DoubleValuesSource {
         + " queryVector="
         + Arrays.toString(queryVector)
         + ")";
-  }
-
-  @Override
-  public boolean isCacheable(LeafReaderContext ctx) {
-    return true;
   }
 }
